Separated dict test, added prediction test after unpickling to test AutoFlow.
